pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.3",
    "werkzeug>=3.1.3",
]
